--- a/packetio/buffer_test.go
+++ b/packetio/buffer_test.go
@@ -1,6 +1,7 @@
 package packetio
 
 import (
+	"context"
 	"errors"
 	"fmt"
 	"io"
@@ -17,71 +18,103 @@ func TestBuffer(t *testing.T) {
 	buffer := NewBuffer()
 	packet := make([]byte, 4)
 
-	// Write once
-	n, err := buffer.Write([]byte{0, 1})
-	assert.NoError(err)
-	assert.Equal(2, n)
+	t.Run("Write and read once", func(t *testing.T) {
+		n, err := buffer.Write([]byte{0, 1})
+		assert.NoError(err)
+		assert.Equal(2, n)
 
-	// Read once
-	n, err = buffer.Read(packet)
-	assert.NoError(err)
-	assert.Equal(2, n)
-	assert.Equal([]byte{0, 1}, packet[:n])
+		n, err = buffer.Read(packet)
+		assert.NoError(err)
+		assert.Equal(2, n)
+		assert.Equal([]byte{0, 1}, packet[:n])
+	})
 
-	// Read deadline
-	err = buffer.SetReadDeadline(time.Unix(0, 1))
-	assert.NoError(err)
-	n, err = buffer.Read(packet)
-	if e, ok := err.(net.Error); !ok || !e.Timeout() {
-		t.Errorf("Unexpected error: %v", err)
-	}
-	assert.Equal(0, n)
+	t.Run("Write and read once with context", func(t *testing.T) {
+		ctx := context.Background()
+		n, err := buffer.WriteContext(ctx, []byte{0, 1})
+		assert.NoError(err)
+		assert.Equal(2, n)
 
-	// Reset deadline
-	err = buffer.SetReadDeadline(time.Time{})
-	assert.NoError(err)
+		n, err = buffer.ReadContext(ctx, packet)
+		assert.NoError(err)
+		assert.Equal(2, n)
+		assert.Equal([]byte{0, 1}, packet[:n])
+	})
 
-	// Write twice
-	n, err = buffer.Write([]byte{2, 3, 4})
-	assert.NoError(err)
-	assert.Equal(3, n)
+	t.Run("Read deadline", func(t *testing.T) {
+		err := buffer.SetReadDeadline(time.Unix(0, 1))
+		assert.NoError(err)
+		n, err := buffer.Read(packet)
+		if e, ok := err.(net.Error); !ok || !e.Timeout() {
+			t.Errorf("Unexpected error: %v", err)
+		}
+		assert.Equal(0, n)
+	})
 
-	n, err = buffer.Write([]byte{5, 6, 7})
-	assert.NoError(err)
-	assert.Equal(3, n)
+	t.Run("Read with context", func(t *testing.T) {
+		ctx, cancel := context.WithCancel(context.Background())
+		cancel()
 
-	// Read twice
-	n, err = buffer.Read(packet)
-	assert.NoError(err)
-	assert.Equal(3, n)
-	assert.Equal([]byte{2, 3, 4}, packet[:n])
+		n, err := buffer.ReadContext(ctx, packet)
+		if e, ok := err.(net.Error); !ok || !e.Timeout() {
+			t.Errorf("Unexpected error: %v", err)
+		}
+		assert.Equal(0, n)
+	})
 
-	n, err = buffer.Read(packet)
-	assert.NoError(err)
-	assert.Equal(3, n)
-	assert.Equal([]byte{5, 6, 7}, packet[:n])
+	t.Run("Reset deadline", func(t *testing.T) {
+		err := buffer.SetReadDeadline(time.Time{})
+		assert.NoError(err)
+	})
 
-	// Write once prior to close.
-	_, err = buffer.Write([]byte{3})
-	assert.NoError(err)
+	t.Run("Write twice", func(t *testing.T) {
+		n, err := buffer.Write([]byte{2, 3, 4})
+		assert.NoError(err)
+		assert.Equal(3, n)
 
-	// Close
-	err = buffer.Close()
-	assert.NoError(err)
+		n, err = buffer.Write([]byte{5, 6, 7})
+		assert.NoError(err)
+		assert.Equal(3, n)
+	})
 
-	// Future writes will error
-	_, err = buffer.Write([]byte{4})
-	assert.Error(err)
+	t.Run("Read twice", func(t *testing.T) {
+		n, err := buffer.Read(packet)
+		assert.NoError(err)
+		assert.Equal(3, n)
+		assert.Equal([]byte{2, 3, 4}, packet[:n])
 
-	// But we can read the remaining data.
-	n, err = buffer.Read(packet)
-	assert.NoError(err)
-	assert.Equal(1, n)
-	assert.Equal([]byte{3}, packet[:n])
+		n, err = buffer.Read(packet)
+		assert.NoError(err)
+		assert.Equal(3, n)
+		assert.Equal([]byte{5, 6, 7}, packet[:n])
+	})
 
-	// Until EOF
-	_, err = buffer.Read(packet)
-	assert.Equal(io.EOF, err)
+	t.Run("Write once prior to close", func(t *testing.T) {
+		_, err := buffer.Write([]byte{3})
+		assert.NoError(err)
+	})
+
+	t.Run("Close", func(t *testing.T) {
+		err := buffer.Close()
+		assert.NoError(err)
+	})
+
+	t.Run("Future writes will error", func(t *testing.T) {
+		_, err := buffer.Write([]byte{4})
+		assert.Error(err)
+	})
+
+	t.Run("But we can read the remaining data", func(t *testing.T) {
+		n, err := buffer.Read(packet)
+		assert.NoError(err)
+		assert.Equal(1, n)
+		assert.Equal([]byte{3}, packet[:n])
+	})
+
+	t.Run("Until EOF", func(t *testing.T) {
+		_, err := buffer.Read(packet)
+		assert.Equal(io.EOF, err)
+	})
 }
 
 func testWraparound(t *testing.T, grow bool) {
@@ -478,19 +511,41 @@ func benchmarkBufferWR(b *testing.B, size int64, write bool, grow int) { // noli
 		}
 	}
 
-	b.SetBytes(size)
-	b.ResetTimer()
+	b.Run("NoContext", func(b *testing.B) {
+		b.ResetTimer()
+		b.SetBytes(size)
 
-	for i := 0; i < b.N; i++ {
-		_, err := buffer.Write(packet)
-		if err != nil {
-			b.Fatalf("Write: %v", err)
+		for i := 0; i < b.N; i++ {
+			_, err := buffer.Write(packet)
+			if err != nil {
+				b.Fatalf("Write: %v", err)
+			}
+			_, err = buffer.Read(packet)
+			if err != nil {
+				b.Fatalf("Read: %v", err)
+			}
 		}
-		_, err = buffer.Read(packet)
-		if err != nil {
-			b.Fatalf("Read: %v", err)
+		b.StopTimer()
+	})
+	b.Run("Context", func(b *testing.B) {
+		ctx, cancel := context.WithCancel(context.Background())
+		defer cancel()
+
+		b.ResetTimer()
+		b.SetBytes(size)
+
+		for i := 0; i < b.N; i++ {
+			_, err := buffer.WriteContext(ctx, packet)
+			if err != nil {
+				b.Fatalf("Write: %v", err)
+			}
+			_, err = buffer.ReadContext(ctx, packet)
+			if err != nil {
+				b.Fatalf("Read: %v", err)
+			}
 		}
-	}
+		b.StopTimer()
+	})
 }
 
 // In this benchmark, the buffer is often empty, which is hopefully