@Sean-Der 2020-11-28
I don't understand how without Context can be slower. Is `context.Background()` every time expensive, could we have a stored one on the buffer?

Reading the benchmark now.