@at-wat 2020-11-28
In my understanding, reading channel internally uses futex syscall which is relatively slow.
So, adding more channels to select syntax increases overhead more.