@Sean-Der 2020-11-28
How does it look when you drop the Read Deadline? Does anyone even use that today.

We might be able to just go `Context` only and do `github.com/pion/transport/v2`. We don't expose the buffer in `pion/webrtc` so I don't feel to bad breaking it!