@at-wat 2020-11-28
I think Read with deadline would be better to exist to make it compatible with Go1's `io.Reader`. (not very strong opinion)
As most of the code is shared between them, it doesn't requires much effort to maintain.
We can drop it if it costs too much to be maintained though.