@at-wat 2020-11-28
Opened an issue for the ordinal Read/Write interface.